+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.3] - 2019-02-27
+
+- silent flush/push error handling if user does not provide handlers
+- hard `maxLength` parameter added to enforce an absolute maximum queue length
+
 ## [0.1.2] - 2019-02-26
 
 - fix error message and add stack trace to error handling